pyproject.toml
[tool.poetry]
name = "hw06"
version = "0.1.0"
description = ""
authors = ["AnybodyKnows <evgeniiolavrik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.9"
faker = "^24.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
